refactor(ward): update api
