+++ src/hospital/hospital.service.ts
@@ -10,15 +10,15 @@ import { CreateHospitalDto } from './dto/request-dto/create-hospital.dto';
 import { Hospital } from './entities/hospital.entity';
 import { hash } from 'bcrypt';
 import { IdCheckDto } from './dto/request-dto/id-check.dto';
-import { CreateWardDto } from './dto/request-dto/create-ward.dto';
+import { CreateWardDto } from '../ward/dto/request-dto/create-ward.dto';
 import { Ward } from '../ward/entities/ward.entity';
 import { CreateRoomDto } from './dto/request-dto/create-room.dto';
 import { Room } from './entities/room.entity';
 import { Patient } from './entities/patient.entity';
 import { CreatePatientDto } from './dto/request-dto/create-patient.dto';
 import { Reservation } from '../reservation/entities/reservation.entity';
-import { UpdateWardDto } from './dto/request-dto/update-ward.dto';
-import { DeleteWardDto } from './dto/request-dto/delete-ward.dto';
+import { UpdateWardDto } from '../ward/dto/request-dto/update-ward.dto';
+import { DeleteWardDto } from '../ward/dto/request-dto/delete-ward.dto';
 import { UpdateRoomDto } from './dto/request-dto/update-room.dto';
 import { DeleteRoomDto } from './dto/request-dto/delete-room.dto';
 import { UpdatePatientDto } from './dto/request-dto/update-patient.dto';
@@ -417,34 +417,6 @@ export class HospitalService {
       }
     }
     return result;
-  }
-
-  async updateWard(requestDto: UpdateWardDto, hospitalId: string) {
-    const hospital = await this.findHospital(hospitalId);
-
-    const ward = await this.wardRepository
-      .createQueryBuilder('ward')
-      .select('ward')
-      .where('ward.id =:id', { id: requestDto.id })
-      .andWhere('ward.hospitalId =:hospitalId', { hospitalId: hospital.id })
-      .execute();
-
-    if (ward.length != 1) {
-      throw new BadRequestException({
-        statusCode: HttpStatus.BAD_REQUEST,
-        message: ['병원과 병동 정보가 올바르지 않습니다.'],
-        error: 'BAD_REQUEST',
-      });
-    }
-
-    const result = await this.wardRepository.update(
-      { id: requestDto.id },
-      { name: requestDto.name },
-    );
-
-    if (result.affected > 0) {
-      return requestDto;
-    }
   }
 
   async updateRoom(hospitalId: string, requestDto: UpdateRoomDto) {
