+++ src/ward/ward.controller.ts
@@ -4,6 +4,7 @@ import {
   Delete,
   HttpStatus,
   Post,
+  Put,
   Req,
   Res,
   UseGuards,
@@ -14,12 +15,14 @@ import {
   ApiOperation,
 } from '@nestjs/swagger';
 import { JwtAuthGuard } from 'src/auth/guards/jwt-auth.guard';
-import { DeleteWardDto } from 'src/hospital/dto/request-dto/delete-ward.dto';
+import { DeleteWardDto } from 'src/ward/dto/request-dto/delete-ward.dto';
 import { BaseResponse } from 'src/util/swagger/base-response.dto';
 import { Request, Response } from 'express';
 import { WardService } from './ward.service';
-import { CreateWardResponse } from 'src/hospital/dto/response-dto/create-ward-response.dto';
-import { CreateWardDto } from 'src/hospital/dto/request-dto/create-ward.dto';
+import { CreateWardResponse } from 'src/ward/dto/response-dto/create-ward-response.dto';
+import { CreateWardDto } from 'src/ward/dto/request-dto/create-ward.dto';
+import { UpdateWardResponse } from 'src/ward/dto/response-dto/update-ward-response.dto';
+import { UpdateWardDto } from 'src/ward/dto/request-dto/update-ward.dto';
 
 @Controller('')
 export class WardController {
@@ -50,6 +53,33 @@ export class WardController {
       ward: ward,
     };
     return res.status(HttpStatus.CREATED).json(result);
+  }
+
+  @Put('hospital/ward')
+  @UseGuards(JwtAuthGuard)
+  @ApiOperation({
+    summary: '병동 수정',
+    description: '병동 수정',
+  })
+  @ApiOkResponse({
+    status: HttpStatus.OK,
+    description: 'success',
+    type: UpdateWardResponse,
+  })
+  async updateWard(
+    @Body() requestDto: UpdateWardDto,
+    @Req() req: Request,
+    @Res() res: Response,
+  ) {
+    const ward = await this.wardService.updateWard(
+      requestDto,
+      req.user.hospitalId,
+    );
+    const result = {
+      message: 'success',
+      ward: ward,
+    };
+    return res.status(HttpStatus.OK).json(result);
   }
 
   @Delete('hospital/ward')
